fix visual studio 2010 build
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@6903 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: b85a78f2ab0463dcbe134636f70878627eab542a
Former-commit-id: 7866d0a11b286dc1dc45becb4e72557dbe4354a5

+++ src/modules/userinterface/legacymenu/racescreens/carsetupmenu.cpp
@@ -280,9 +280,9 @@ void CarSetupMenu::loadSettings()
 
             attnum &att = items[page][index];
 
-            att.labelId = getDynamicControlId(std::string("Label" + std::to_string(index)).c_str());
-            att.editId = getDynamicControlId(std::string("Edit" + std::to_string(index)).c_str());
-            att.defaultLabelId = getDynamicControlId(std::string("DefaultLabel" + std::to_string(index)).c_str());
+            att.labelId = getDynamicControlId(std::string("Label" + std::to_string(static_cast<unsigned long long>(index))).c_str());
+            att.editId = getDynamicControlId(std::string("Edit" + std::to_string(static_cast<unsigned long long>(index))).c_str());
+            att.defaultLabelId = getDynamicControlId(std::string("DefaultLabel" + std::to_string(static_cast<unsigned long long>(index))).c_str());
             att.section = GfParmGetStr(hparmItems, strSection.c_str(), "section", "");
             att.param = GfParmGetStr(hparmItems, strSection.c_str(), "param", "");
             att.units = GfParmGetStr(hparmItems, strSection.c_str(), "unit", "");
@@ -316,11 +316,11 @@ void CarSetupMenu::loadSettings()
                 attnum &att = items[page][index];
 
                 if (!att.labelId)
-                    att.labelId = getDynamicControlId(std::string("Label" + std::to_string(index)).c_str());
+                    att.labelId = getDynamicControlId(std::string("Label" + std::to_string(static_cast<unsigned long long>(index))).c_str());
                 if (!att.editId)
-                    att.editId = getDynamicControlId(std::string("Edit" + std::to_string(index)).c_str());
+                    att.editId = getDynamicControlId(std::string("Edit" + std::to_string(static_cast<unsigned long long>(index))).c_str());
                 if (!att.defaultLabelId)
-                    att.defaultLabelId = getDynamicControlId(std::string("DefaultLabel" + std::to_string(index)).c_str());
+                    att.defaultLabelId = getDynamicControlId(std::string("DefaultLabel" + std::to_string(static_cast<unsigned long long>(index))).c_str());
             }
         }
     }
@@ -412,9 +412,9 @@ bool CarSetupMenu::initialize(void *pMenu, const GfRace *pRace, const GfDriver *
     // Create items.
     for (size_t index = 0; index < ITEMS_PER_PAGE; ++index)
     {
-        createLabelControl(std::string("Label" + std::to_string(index)).c_str());
-        createEditControl(std::string("Edit" + std::to_string(index)).c_str(), this, NULL, NULL);
-        createLabelControl(std::string("DefaultLabel" + std::to_string(index)).c_str());
+        createLabelControl(std::string("Label" + std::to_string(static_cast<unsigned long long>(index))).c_str());
+        createEditControl(std::string("Edit" + std::to_string(static_cast<unsigned long long>(index))).c_str(), this, NULL, NULL);
+        createLabelControl(std::string("DefaultLabel" + std::to_string(static_cast<unsigned long long>(index))).c_str());
     }
 
     // Create buttons.
